Use OpenTelemetry for the Function Worker only, instead of the spawned function processes too
