+++ pulsar-functions/instance/src/main/java/org/apache/pulsar/functions/instance/ContextImpl.java
@@ -69,7 +69,6 @@ import org.apache.pulsar.functions.instance.state.StateManager;
 import org.apache.pulsar.functions.instance.stats.ComponentStatsManager;
 import org.apache.pulsar.functions.instance.stats.FunctionCollectorRegistry;
 import org.apache.pulsar.functions.instance.stats.FunctionStatsManager;
-import org.apache.pulsar.functions.instance.stats.PulsarWorkerOpenTelemetry;
 import org.apache.pulsar.functions.instance.stats.SinkStatsManager;
 import org.apache.pulsar.functions.instance.stats.SourceStatsManager;
 import org.apache.pulsar.functions.proto.Function;
@@ -119,7 +118,6 @@ class ContextImpl implements Context, SinkContext, SourceContext, AutoCloseable 
     Map<String, String[]> userMetricsLabels = new HashMap<>();
     private final String[] metricsLabels;
     private final Summary userMetricsSummary;
-    private final PulsarWorkerOpenTelemetry openTelemetry;
 
     private final SubscriptionType subscriptionType;
 
@@ -225,7 +223,6 @@ class ContextImpl implements Context, SinkContext, SourceContext, AutoCloseable 
                         .quantile(0.99, 0.01)
                         .quantile(0.999, 0.01)
                         .create());
-        this.openTelemetry = new PulsarWorkerOpenTelemetry(config);
         this.componentType = componentType;
         this.stateManager = stateManager;
         this.defaultStateStore = (DefaultStateStore) stateManager.getStore(
@@ -756,10 +753,6 @@ class ContextImpl implements Context, SinkContext, SourceContext, AutoCloseable 
             CompletableFuture.allOf(futures.toArray(new CompletableFuture[0])).get();
         } catch (InterruptedException | ExecutionException e) {
             logger.warn("Failed to close producers", e);
-        }
-
-        if (openTelemetry != null) {
-            openTelemetry.close();
         }
     }
